refactor(packs): extract ExecutionContext and result types
Pulls the dependency-injection envelope and the result/Command types
out of orchestration.rs into focused submodules:

- packs/context.rs — ExecutionContext struct + production() constructor
- packs/types.rs   — PackResult, ExecuteResult, Command trait

orchestration.rs keeps the actual pipeline (execute, plan_pack,
filter_pre_preprocess_gates, …) and re-exports the moved items so the
existing crate::packs::orchestration::ExecutionContext path keeps
working — no callers need to change.
