+++ crates/dodot-lib/src/packs/orchestration.rs
@@ -3,180 +3,26 @@
 //!
 //! `execute()` owns the outer loop: discover packs → load per-pack
 //! config → execute command → aggregate results.
+//!
+//! The shared [`ExecutionContext`] (every command's dependency bag) and
+//! the result types ([`PackResult`], [`ExecuteResult`]) plus the
+//! per-pack [`Command`] trait live in sibling modules
+//! ([`crate::packs::context`], [`crate::packs::types`]); they are
+//! re-exported here for the historical `crate::packs::orchestration::X`
+//! surface.
 
-use std::sync::Arc;
-
-use serde::Serialize;
 use tracing::{debug, info};
 
-use crate::config::ConfigManager;
-use crate::datastore::DataStore;
 use crate::execution::Executor;
-use crate::fs::Fs;
 use crate::gates::{GateTable, HostFacts};
 use crate::handlers;
 use crate::operations::OperationResult;
 use crate::packs::{self, Pack};
-use crate::paths::Pather;
 use crate::rules::{self, Scanner};
 use crate::Result;
 
-// ── Types ───────────────────────────────────────────────────────
-
-/// Everything the pipeline needs to execute.
-pub struct ExecutionContext {
-    pub fs: Arc<dyn Fs>,
-    pub datastore: Arc<dyn DataStore>,
-    pub paths: Arc<dyn Pather>,
-    pub config_manager: Arc<ConfigManager>,
-    /// Pre-flight syntax checker for shell sources. Production wires
-    /// up [`SystemSyntaxChecker`](crate::shell::SystemSyntaxChecker)
-    /// (spawns real `bash`/`zsh -n`); tests inject a mock.
-    pub syntax_checker: Arc<dyn crate::shell::SyntaxChecker>,
-    /// Subprocess runner for advisory probes (homebrew-cask lookup,
-    /// macOS `mdls`/`mdfind`). Production reuses the same
-    /// [`ShellCommandRunner`](crate::datastore::ShellCommandRunner)
-    /// the datastore uses for handler-driven commands; tests inject a
-    /// mock that returns canned outputs without spawning processes.
-    /// See `docs/proposals/macos-paths.lex` §8.
-    pub command_runner: Arc<dyn crate::datastore::CommandRunner>,
-    pub dry_run: bool,
-    pub no_provision: bool,
-    pub provision_rerun: bool,
-    pub force: bool,
-    /// How pack-status output should render rows: `Full` keeps today's
-    /// per-file listing, `Short` collapses each pack to one summary
-    /// line. Consumed by every command that renders through the
-    /// `pack-status` template (`status`, `up`, `down`, `adopt`);
-    /// ignored by commands that emit `message` / `list` output.
-    pub view_mode: crate::commands::ViewMode,
-    /// How packs are ordered in pack-status output: `Name` (flat
-    /// alphabetical / discovery order) or `Status` (grouped under
-    /// Ignored / Deployed / Pending / Error banners). Consumed by
-    /// every command that renders through the `pack-status` template;
-    /// ignored by commands that emit `message` / `list` output.
-    pub group_mode: crate::commands::GroupMode,
-    /// When true, install-script execution streams raw stdout/stderr
-    /// to the user's terminal. The default (`false`) keeps output
-    /// quiet — only the `# status:` progress markers and the leading
-    /// comment block of each script are surfaced. Wired from the CLI
-    /// global `--verbose`/`--debug` flag.
-    pub verbose: bool,
-    /// Snapshot of the host's gate-relevant facts (os, arch, hostname,
-    /// username). Detected once per context so per-pack scanning and
-    /// matching avoid re-running `hostname(1)`/env reads. Constructed
-    /// by [`Self::production`]; tests build via `HostFacts::for_tests`.
-    pub host_facts: Arc<HostFacts>,
-}
-
-impl ExecutionContext {
-    /// Create a default production context from a dotfiles root path.
-    ///
-    /// Wires up the real filesystem, XDG paths, filesystem-backed
-    /// datastore with shell command runner, and clapfig config manager.
-    /// `verbose` controls whether install-script stdout/stderr is
-    /// streamed to the terminal; the field is also stored on the
-    /// returned context for any other consumer that cares. Callers
-    /// only need to override specific fields (e.g. `dry_run`).
-    pub fn production(dotfiles_root: &std::path::Path, verbose: bool) -> crate::Result<Self> {
-        let config_manager = Arc::new(ConfigManager::new(dotfiles_root)?);
-
-        // Honor `app_uses_library = false` by collapsing app_support_dir
-        // onto xdg_config_home — that's the "Linux-style ~/.config
-        // everywhere even on macOS" escape hatch from
-        // `docs/proposals/macos-paths.lex` §6.3 / §11.2.
-        //
-        // Soft-fail by design: a config-load failure here only blocks
-        // the `app_uses_library = false` override from being applied,
-        // not context construction itself. Real config errors (parse
-        // failures, missing required fields) bubble up the next time a
-        // command calls `config_manager.root_config()` — same surface,
-        // same error path, just without preempting Pather construction.
-        // If the read fails here we leave `app_support_dir` at the
-        // platform default and let the actual command surface the error.
-        let mut paths_builder = crate::paths::XdgPather::builder().dotfiles_root(dotfiles_root);
-        if let Ok(root_config) = config_manager.root_config() {
-            if !root_config.symlink.app_uses_library {
-                // Resolve XDG the way XdgPatherBuilder will, then pin
-                // app_support_dir at the same path. We can't read the
-                // builder's resolved xdg back out before build(), so
-                // duplicate the precedence here.
-                let home = std::env::var("HOME")
-                    .map(std::path::PathBuf::from)
-                    .unwrap_or_else(|_| std::path::PathBuf::from("/tmp/dodot-unknown-home"));
-                let xdg = std::env::var("XDG_CONFIG_HOME")
-                    .map(std::path::PathBuf::from)
-                    .unwrap_or_else(|_| home.join(".config"));
-                paths_builder = paths_builder.app_support_dir(xdg);
-            }
-        }
-        let paths = Arc::new(paths_builder.build()?);
-        let fs: Arc<dyn Fs> = Arc::new(crate::fs::OsFs::new());
-        let runner: Arc<dyn crate::datastore::CommandRunner> =
-            Arc::new(crate::datastore::ShellCommandRunner::new(verbose));
-        let datastore: Arc<dyn DataStore> = Arc::new(crate::datastore::FilesystemDataStore::new(
-            fs.clone(),
-            paths.clone(),
-            runner.clone(),
-        ));
-
-        Ok(Self {
-            fs,
-            datastore,
-            paths,
-            config_manager,
-            syntax_checker: Arc::new(crate::shell::SystemSyntaxChecker),
-            command_runner: runner,
-            dry_run: false,
-            no_provision: false,
-            provision_rerun: false,
-            force: false,
-            view_mode: crate::commands::ViewMode::default(),
-            group_mode: crate::commands::GroupMode::default(),
-            verbose,
-            host_facts: Arc::new(HostFacts::detect()),
-        })
-    }
-}
-
-/// Result for a single pack.
-#[derive(Debug, Serialize)]
-pub struct PackResult {
-    pub pack_name: String,
-    pub success: bool,
-    pub operations: Vec<OperationResult>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub error: Option<String>,
-}
-
-/// Aggregated result across all packs.
-#[derive(Debug, Serialize)]
-pub struct ExecuteResult {
-    pub pack_results: Vec<PackResult>,
-    pub total_packs: usize,
-    pub successful_packs: usize,
-    pub failed_packs: usize,
-}
-
-impl ExecuteResult {
-    pub fn is_success(&self) -> bool {
-        self.failed_packs == 0
-    }
-}
-
-// ── Command trait ───────────────────────────────────────────────
-
-/// A command that operates on a single pack.
-///
-/// The orchestration pipeline calls `execute_for_pack` for each
-/// discovered pack. Commands implement the specific logic (up, down,
-/// status, etc.) while the pipeline handles discovery, config loading,
-/// filtering, and aggregation.
-pub trait Command: Send + Sync {
-    fn name(&self) -> &str;
-
-    fn execute_for_pack(&self, pack: &Pack, ctx: &ExecutionContext) -> Result<PackResult>;
-}
+pub use crate::packs::context::ExecutionContext;
+pub use crate::packs::types::{Command, ExecuteResult, PackResult};
 
 // ── Pipeline ────────────────────────────────────────────────────
 
@@ -1038,9 +884,12 @@ pub fn validate_pack_names(names: &[String], ctx: &ExecutionContext) -> crate::R
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::config::ConfigManager;
     use crate::datastore::{CommandOutput, CommandRunner, FilesystemDataStore};
+    use crate::fs::Fs;
+    use crate::paths::Pather;
     use crate::testing::TempEnvironment;
-    use std::sync::Mutex;
+    use std::sync::{Arc, Mutex};
 
     struct MockCommandRunner {
         calls: Mutex<Vec<String>>,
